+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release notes
 
+## Unreleased
+
+### Feature
+
+- Created `status-badge` component.
 ## 0.9.0
 
 ### Feature