+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Feature
 
 - Created `status-badge` component.
+
 ## 0.9.0
 
 ### Feature